Edit: reservation dir structure - move modals into Modal dir
