+++ src/pages/Reservation/Modal/PayOptionSelector.module.scss
@@ -1,19 +1,13 @@
-@import '../../styles/common.scss';
+@import '../../../styles/common.scss';
 
 .container {
-	position: fixed;
-	left: 30%;
-	top: 40%;
-	-ms-transform: translate(-50%, -50%);
-	-moz-transform: translate(-50%, -50%);
-	-webkit-transform: translate(-50%, -50%);
-	transform: translate(-50%, -50%);
-	z-index: 3;
+	display: flex;
+	position: relative;
+	z-index:2;
 	font-family: $theme-font;
 	color: $theme-font-color;
-
-
-	.reservationNotValidmessage {
+	
+	.dropDown {
 		display: flex;
 		flex-direction: column;
 		align-items: stretch;
@@ -23,49 +17,49 @@
 		box-shadow: $tooltip-shadow;
 		position: absolute;
 		border-radius: 5px;
-
-		.messageHeader {
+		
+		.dropDownHeader {
 			display: flex;
 			border-bottom: $theme-border-line;
-
-			.CloseBtnBox {
-				.CloseBtn {
-					padding: 10px;
+			cursor: auto;
+			.dropdownCloseBtnBox {
+				@include flex-v-center;
+				.dropdownCloseBtn {
+					padding: 0 15px;
 					background-color: white;
 					border: none;
-				}
 
-				svg {
-					font-size: 20px;
+					&:hover {
+						cursor: pointer;
+					}
+
+					svg {
+						font-size: 15px;
+					}
 				}
 			}
-
-			.confirmationFailedHeaderTextBox {
+			.dropdownHeaderTextBox {
 				min-height: 30px;
-				margin-left: 60px;
+				margin-left: 40px;
 				margin-right: 80px;
-
 				p {
-					padding: 60px 40px;
-					@include h2;
-					
+					padding: 15px;
+					@include h3;
 				}
 			}
-
 		}
-
-		.messageContents {
-			display: flex;
+		.dropDownSelectors {
+			display:flex;
 			flex-direction: column;
 			align-items: stretch;
 			align-content: center;
 			padding: 5px 18px 10px 18px;
 
 			div {
+				@include flex-v-center;
 				min-height: 40px;
 				border-bottom: $theme-border-line;
 			}
 		}
-
 	}
 }
