+++ src/pages/Reservation/Modal/PinnedBox.module.scss
