+++ src/pages/Reservation/Modal/PriceBreakDown.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/common.scss';
+@import '../../../styles/common.scss';
 
 .container {
 	@include flex-direction-column;
