+++ src/pages/Reservation/Modal/ReservationConfirmed.module.scss
@@ -1,12 +1,18 @@
-@import '../../styles/common.scss';
+@import '../../../styles/common.scss';
 
 .container {
-	display: flex;
-	position: relative;
+	position: fixed;
+	left: 50%;
+	top: 50%;
+	-ms-transform: translate(-50%, -50%);
+	-moz-transform: translate(-50%, -50%);
+	-webkit-transform: translate(-50%, -50%);
+	transform: translate(-50%, -50%);
+	z-index: 3;
 	font-family: $theme-font;
 	color: $theme-font-color;
-	
-	.dropDown {
+
+	.confirmationMessage {
 		display: flex;
 		flex-direction: column;
 		align-items: stretch;
@@ -16,44 +22,50 @@
 		box-shadow: $tooltip-shadow;
 		position: absolute;
 		border-radius: 5px;
-		
-		.dropDownHeader {
+
+		.messageHeader {
 			display: flex;
 			border-bottom: $theme-border-line;
-			cursor: auto;
-			.dropdownCloseBtnBox {
-				.dropdownCloseBtn {
+
+			.CloseBtnBox {
+				.CloseBtn {
 					padding: 10px;
 					background-color: white;
 					border: none;
-					&:hover {
-						cursor: pointer;
-					}
 				}
+
 				svg {
 					font-size: 15px;
 				}
 			}
-			.dropdownHeaderTextBox {
+
+			.confirmationHeaderTextBox {
 				min-height: 30px;
-				margin-left: 40px;
+				margin-left: 60px;
 				margin-right: 80px;
+
 				p {
 					padding: 15px;
 					@include h3;
+					
 				}
 			}
-			
+
 		}
+
 		.dropDownSelectors {
-			display:flex;
+			display: flex;
 			flex-direction: column;
 			align-items: stretch;
 			align-content: center;
 			padding: 5px 18px 10px 18px;
 
+			div {
+				min-height: 40px;
+				border-bottom: $theme-border-line;
+			}
 		}
-		
+
 	}
