+++ src/pages/Reservation/Reservation.js
@@ -3,11 +3,11 @@ import { useNavigate, useLocation } from 'react-router-dom';
 import Home from '../Home/Home';
 import css from './Reservation.module.scss';
 import Footer from '../../components/Footer/Footer';
-import PayOptionSelector from './PayOptionSelector';
-import CardInfoCountrySelector from './CardInfoCountrySelector';
-import ReservationConfirmed from './ReservationConfirmed';
-import ReservationNotValid from './ReservationNotValid';
-import PinnedBox from './modals/PinnedBox';
+import PayOptionSelector from './Modal/PayOptionSelector';
+import CardInfoCountrySelector from './Modal/CardInfoCountrySelector';
+import ReservationConfirmed from './Modal/ReservationConfirmed';
+import ReservationNotValid from './Modal/ReservationNotValid';
+import PinnedBox from './Modal/PinnedBox';
 import { MdNavigateBefore, MdCreditCard } from 'react-icons/md';
 import { SiVisa, SiMastercard } from 'react-icons/si';
 import { GrAmex } from 'react-icons/gr';
